Record tactic path identifiers rather than string names.
This means that all tactics now have to be rebound in hooks, including ACCEPT_TAC,
since the identifier is not available at the point the tactic is declared.

This means that we cannot so easily separate the hook stuff and the tactic code, so
we'll just merge those files together.

To reduce duplication, we provide install_tactic_transformer. This is still not
great, since in itab.ml, we need to reinstall the previous hook. The hack of the
ocaml compiler currently doesn't allow us to peek at the old set of hooks, so this
could lead to bugs if we forget what the old hook is supposed to be.