Fix for clearing local dups.
NOTE: This requires "prove_subsumed_by" to make sure that a theorem of the original
form is proven. The proof in this function cannot be put in Acc, unless we can make
Acc and the kernel mutually recursive. For now, we just shove in the function on the
assumption that we're using alphaconv for testing for duplicates.