Don't partial apply transform_item.
Before, we built up the parts of the AST relating to "foo" in advance. The problem
here is that a rebinding of "foo" to a different type breaks us out of the type
system. Finding the type of "foo" on every tree transform allows us to catch this
problem and bale safely.